requirements.txt
coreapi==2.2.4
Django==1.10.5
dj-database-url==0.4.2
rest-framework-generic-relations==1.1.0
